requirements.txt
streamlit
pandas
python-dateutil
openai
